requirements.txt
PyQt6>=6.4.0
PyMuPDF>=1.23.0
easyocr>=1.7.0
Pillow>=10.0.0
google-genai>=0.3.0
